requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
google-generativeai==0.3.2
python-multipart==0.0.6
Pillow==10.2.0
